complete refactoring of the build system.
Now the build system create *.a lib for each arch.
Each example can be linked with one of a compiled lib
