+++ README.md
@@ -10,7 +10,7 @@ The role of this code is to manage the Poppy 2.0 communication stack between all
 
 It is the basic library used to create new modules, you can use it to develop you own !
 
-To understand how this protocole work, please read the [protocol definition](doc/protocol_definition.md).
+To understand how this protocole work, please read the [protocol definition](extra/doc/protocol_definition.md).
 
 How to start a new Poppy 2.0 module code project
 ------------------------------------------------
@@ -23,7 +23,7 @@ __________________________________________________________________________
 In this repo travis is used to create documentation, update readme, and pass some tests.
 To activate all this things you have to follow some steps :
  - enable your repo on [travis](https://travis-ci.org/) and on [coveralls](https://coveralls.io).
- - [Create an encrypted token for travis.](doc/travis_encrypt.md)
+ - [Create an encrypted token for travis.](extra/doc/travis_encrypt.md)
  - replace your encrypted token to the .travis.yml file of this repo.
 
 __________________________________________________________________________
@@ -33,7 +33,7 @@ To have a perfect integration and declaration of your module into a poppy 2.0 ne
  - specify your MCU on the Makefile.
  - specify your MAINCLOCK frequency on the Makefile.
  - specify your SCLFREQ to specify the I2C max speed on the Makefile.
- - create your specific hal folder if you need. To do that please read the [new hal creation documentation](doc/hal_creation.md).
+ - create your specific hal folder if you need. To do that please read the [new hal creation documentation](extra/doc/hal_creation.md).
 
 You can use the [template.c](template.c) file to format your main file and use it as base.
