+++ extra/examples/template/Makefile
@@ -1,0 +1,11 @@
+#TODO faire une entête avec licence et tout
+
+SHELL := /bin/sh
+
+.SUFFIXES: .d .o .c .h .cpp .hpp .s .S
+
+PROJECT_NAME := template
+
+include ../preambule.mk
+
+include ../rules.mk
