+++ extra/examples/tests/inc/test_mngmnt.h
