+++ poppy-com/hal/arduino_uno/mcu.mk
@@ -1,0 +1,9 @@
+#-------------------------------------------------------------------------------
+#       Variant information
+#-------------------------------------------------------------------------------
+
+BOARD = arduino_uno
+MCU = atmega328p
+DEVICE_CORE = AVR
+CROSS_COMPILE = avr-
+MAINCLOCK = 16000000ul
