+++ poppy-com/hal/atmega64/mcu.mk
@@ -1,0 +1,9 @@
+#-------------------------------------------------------------------------------
+#       Variant information
+#-------------------------------------------------------------------------------
+
+BOARD = atmega64
+MCU = atmega64
+DEVICE_CORE = AVR
+CROSS_COMPILE = avr-
+MAINCLOCK = 16000000ul
