+++ poppy-com/hal/stub/Makefile
@@ -1,0 +1,8 @@
+#TODO faire une entête avec licence et tout
+
+SHELL := /bin/sh
+
+.SUFFIXES: .d .o .c .h .cpp .hpp .s .S
+
+include mcu.mk
+include ../rules.mk
