+++ poppy-com/hal/stub/mcu.mk
@@ -1,0 +1,9 @@
+#-------------------------------------------------------------------------------
+#       Variant information
+#-------------------------------------------------------------------------------
+
+BOARD = stub
+MCU = x86
+DEVICE_CORE = x86
+CROSS_COMPILE =
+MAINCLOCK = unknown
