+++ poppy-com/hal/template/mcu.mk
@@ -1,0 +1,9 @@
+#-------------------------------------------------------------------------------
+#       Variant information
+#-------------------------------------------------------------------------------
+
+BOARD = template
+MCU = device_name
+DEVICE_CORE = core_type (AVR, ARM, X86)
+CROSS_COMPILE = "avr-" or ""
+MAINCLOCK = example 120000000ul for 120Mhz
