+++ poppy-com/inc/i2c_master.h
@@ -1,7 +1,7 @@
 #ifndef I2C_MASTER_H_
 #define I2C_MASTER_H_
 
-#include "poppy-com/inc/context.h"
+#include "inc/context.h"
 
 unsigned char set_extern_id(unsigned char addr, unsigned char newid);
 unsigned char get_extern_module_type(unsigned char addr, unsigned char *module_type);
